+++ frontend-client/lib/utils.ts
@@ -1,8 +1,6 @@
 import { type ClassValue, clsx } from 'clsx'
 import ky from 'ky'
-import { getServerSession } from 'next-auth'
 import { twMerge } from 'tailwind-merge'
-import { authOptions } from './auth'
 import { baseUrl } from './vars'
 
 export const cn = (...inputs: ClassValue[]) => {
@@ -11,19 +9,5 @@ export const cn = (...inputs: ClassValue[]) => {
 
 export const fetcher = ky.create({
   prefixUrl: baseUrl,
-  credentials: 'include'
-})
-
-// TODO: add refresh token logic
-export const fetcherWithAuth = fetcher.extend({
-  hooks: {
-    beforeRequest: [
-      async (request) => {
-        const session = await getServerSession(authOptions)
-        if (session) {
-          request.headers.set('Authorization', session.user.accessToken)
-        }
-      }
-    ]
-  }
+  throwHttpErrors: false
 })
