+++ frontend-client/lib/vars.ts
@@ -1,1 +1,9 @@
 export const baseUrl = process.env.NEXT_PUBLIC_BASEURL
+
+const SECONDS_PER_MINUTE = 60
+const SECONDS_PER_HOUR = 60 * SECONDS_PER_MINUTE
+const SECONDS_PER_DAY = 24 * SECONDS_PER_HOUR
+
+/** JWT Token Expiration Settings */
+export const ACCESS_TOKEN_EXPIRE_TIME = 30 * SECONDS_PER_MINUTE
+export const REFRESH_TOKEN_EXPIRE_TIME = SECONDS_PER_DAY
