+++ frontend-client/types/next-auth.d.ts
@@ -1,28 +1,34 @@
 import type { DefaultSession, DefaultUser } from 'next-auth'
 
 declare module 'next-auth' {
-  interface Session extends DefaultSession {
-    user: {
-      username: string
-      role: string
-      accessToken: string
-      refreshToken: string
-    }
-  }
-
   interface User extends DefaultUser {
     username: string
     role: string
     accessToken: string
     refreshToken: string
+    accessTokenExpires: number
+    refreshTokenExpires: number
+  }
+  interface Session extends DefaultSession {
+    user: {
+      username: string
+      role: string
+    }
+    token: {
+      accessToken: string
+      refreshToken: string
+      accessTokenExpires: number
+      refreshTokenExpires: number
+    }
   }
 }
-
 declare module 'next-auth/jwt' {
   interface JWT {
     username: string
     role: string
     accessToken: string
     refreshToken: string
+    accessTokenExpires: number
+    refreshTokenExpires: number
   }
 }
